Project
Unfortunate events have occurred during the last few days, and the new intern
at IT department managed to leave a backdoor open. The hackers have taken
advantage of this situation and have managed to infiltrate the system. The
hackers have left a message in the system, but it is gibirish and the IT department
needs your help to decrypt it. The message left by the hackers can contain any
characters, including but not limited to spaces, numbers, and special characters,
but it includes 2 important words ZERO and ONE.
In order to understand the message, you need to decrypt it. There are a few
steps to decrypt the message:
1. Extract the occurrence of the word ZERO and ONE from the given string
and put them to an array.
2. Transfer the extracted array to a new array with the following rules:
• If the word is ZERO, replace it with 0.
• If the word is ONE, replace it with 1.
3. The output of the previous code is still meaningless. To advance with the
decryption, translate the binary numbers to decimal numbers. Every 8 bit
into a byte.
4. Now we have the decimal numbers, we need to convert them to ASCII
characters.
5. Finally, put all the characters together to get the final result as a string.
6. we came pretty far with the decryption. But we still have a problem. no
body knows the leet language!The B@MER sccientists have provided us
some information that you need for the leet language. You only need to look
for the folowing characters: A, B ,E ,G ,I ,O ,Z They also provided
more information to check the leet language.